moved getCurrentRoute to Route service
